Failed input cancels worker context
If job fails for an input it was canceling worker context, so no other
input could be processed by this worker. So, if we would have 1 worker
in RunConcurrent with 100 inputs and first will fail, no other input
would be processed because of cancelled context.
Fix removed a "global" worker context and replaced it with an input context,
this way if input job fails, it won't continue with other jobs in step but
won't block other inputs from being processed.